requirements.txt
git+https://github.com/opencopilotdev/opencopilot.git@lightyear-copilot
langchain
yfinance==0.2.28
newsapi-python==0.2.7
python-dotenv==1.0.0